requirements.txt
requests==2.31.0
python-dotenv==1.0.1
websockets==12.0
discord.py==2.3.2
